fix: update method names for consistency and improve JSON handling in serialization
