+++ src/cyvest/dsl.py
@@ -255,236 +255,3 @@ class ContainerHandler:
             The container
         """
         return self._container
-
-
-class CyvestDSL:
-    """
-    DSL extension for Cyvest with fluent chaining methods.
-
-    This class extends the base Cyvest functionality with convenient
-    fluent API methods.
-    """
-
-    def __init__(self, cv: Cyvest) -> None:
-        """
-        Initialize the DSL.
-
-        Args:
-            cv: The Cyvest instance to wrap
-        """
-        self._cv = cv
-
-    def observable(
-        self,
-        obs_type: str,
-        value: str,
-        internal: bool = True,
-        whitelisted: bool = False,
-        comment: str = "",
-        extra: dict[str, Any] | None = None,
-        score: Decimal | float | None = None,
-        level: Level | None = None,
-    ) -> ObservableHandler:
-        """
-        Create an observable with fluent interface.
-
-        Args:
-            obs_type: Type of observable
-            value: Value of the observable
-            internal: Whether this is an internal asset
-            whitelisted: Whether this is whitelisted
-            comment: Optional comment
-            extra: Optional extra data
-            score: Optional explicit score
-            level: Optional explicit level
-
-        Returns:
-            Observable handler for chaining
-        """
-        obs = self._cv.observable_create(obs_type, value, internal, whitelisted, comment, extra, score, level)
-        return ObservableHandler(self._cv, obs)
-
-    def create_observable(
-        self,
-        obs_type: str,
-        value: str,
-        internal: bool = True,
-        whitelisted: bool = False,
-        comment: str = "",
-        extra: dict[str, Any] | None = None,
-        score: Decimal | float | None = None,
-        level: Level | None = None,
-    ) -> Observable:
-        """
-        Create an observable (returns the object, not handler).
-
-        Args:
-            obs_type: Type of observable
-            value: Value of the observable
-            internal: Whether this is an internal asset
-            whitelisted: Whether this is whitelisted
-            comment: Optional comment
-            extra: Optional extra data
-            score: Optional explicit score
-            level: Optional explicit level
-
-        Returns:
-            The created observable
-        """
-        return self._cv.observable_create(obs_type, value, internal, whitelisted, comment, extra, score, level)
-
-    def check(
-        self,
-        check_id: str,
-        scope: str,
-        description: str,
-        comment: str = "",
-        extra: dict[str, Any] | None = None,
-        score: Decimal | float | None = None,
-        level: Level | None = None,
-    ) -> CheckHandler:
-        """
-        Create a check with fluent interface.
-
-        Args:
-            check_id: Check identifier
-            scope: Check scope
-            description: Check description
-            comment: Optional comment
-            extra: Optional extra data
-            score: Optional explicit score
-            level: Optional explicit level
-
-        Returns:
-            Check handler for chaining
-        """
-        chk = self._cv.check_create(check_id, scope, description, comment, extra, score, level)
-        return CheckHandler(self._cv, chk)
-
-    def create_check(
-        self,
-        check_id: str,
-        scope: str,
-        description: str,
-        comment: str = "",
-        extra: dict[str, Any] | None = None,
-        score: Decimal | float | None = None,
-        level: Level | None = None,
-    ) -> Check:
-        """
-        Create a check (returns the object, not handler).
-
-        Args:
-            check_id: Check identifier
-            scope: Check scope
-            description: Check description
-            comment: Optional comment
-            extra: Optional extra data
-            score: Optional explicit score
-            level: Optional explicit level
-
-        Returns:
-            The created check
-        """
-        return self._cv.check_create(check_id, scope, description, comment, extra, score, level)
-
-    def container(self, path: str, description: str = "") -> ContainerHandler:
-        """
-        Create a container with fluent interface.
-
-        Args:
-            path: Container path
-            description: Container description
-
-        Returns:
-            Container handler for chaining
-        """
-        ctr = self._cv.container_create(path, description)
-        return ContainerHandler(self._cv, ctr)
-
-    def root(self) -> Observable:
-        """
-        Get the root observable.
-
-        Returns:
-            Root observable
-        """
-        return self._cv.observable_get_root()
-
-
-# Monkey-patch Cyvest to add DSL methods
-def _extend_cyvest() -> None:
-    """Extend Cyvest class with DSL methods."""
-    from cyvest.cyvest import Cyvest
-
-    # Add DSL methods to Cyvest
-    def observable(
-        self: Cyvest,
-        obs_type: str,
-        value: str,
-        internal: bool = True,
-        whitelisted: bool = False,
-        comment: str = "",
-        extra: dict[str, Any] | None = None,
-        score: Decimal | float | None = None,
-        level: Level | None = None,
-    ) -> ObservableHandler:
-        obs = self.observable_create(obs_type, value, internal, whitelisted, comment, extra, score, level)
-        return ObservableHandler(self, obs)
-
-    def create_observable(
-        self: Cyvest,
-        obs_type: str,
-        value: str,
-        internal: bool = True,
-        whitelisted: bool = False,
-        comment: str = "",
-        extra: dict[str, Any] | None = None,
-        score: Decimal | float | None = None,
-        level: Level | None = None,
-    ) -> Observable:
-        return self.observable_create(obs_type, value, internal, whitelisted, comment, extra, score, level)
-
-    def check(
-        self: Cyvest,
-        check_id: str,
-        scope: str,
-        description: str,
-        comment: str = "",
-        extra: dict[str, Any] | None = None,
-        score: Decimal | float | None = None,
-        level: Level | None = None,
-    ) -> CheckHandler:
-        chk = self.check_create(check_id, scope, description, comment, extra, score, level)
-        return CheckHandler(self, chk)
-
-    def create_check(
-        self: Cyvest,
-        check_id: str,
-        scope: str,
-        description: str,
-        comment: str = "",
-        extra: dict[str, Any] | None = None,
-        score: Decimal | float | None = None,
-        level: Level | None = None,
-    ) -> Check:
-        return self.check_create(check_id, scope, description, comment, extra, score, level)
-
-    def container(self: Cyvest, path: str, description: str = "") -> ContainerHandler:
-        ctr = self.container_create(path, description)
-        return ContainerHandler(self, ctr)
-
-    def root(self: Cyvest) -> Observable:
-        return self.observable_get_root()
-
-    # Apply monkey patches
-    Cyvest.observable = observable  # type: ignore
-    Cyvest.create_observable = create_observable  # type: ignore
-    Cyvest.check = check  # type: ignore
-    Cyvest.create_check = create_check  # type: ignore
-    Cyvest.container = container  # type: ignore
-    Cyvest.root = root  # type: ignore
-
-
-# Auto-extend on import
-_extend_cyvest()
